# packages/kanban-viewer/openapi.yaml
openapi: 3.1.0
info:
  title: A(i)-Team Kanban Viewer API
  version: 1.0.0
  description: |
    REST API for the A(i)-Team Kanban Viewer. Manages work items, missions,
    agent lifecycle, and real-time board events for the TDD pipeline.

    ## Project Isolation
    All requests (except `/api/projects`) require the `X-Project-ID` header.
    Projects are created automatically on first use via `POST /api/projects`.

    ## Error Format
    All errors return a consistent envelope:
    ```json
    { "success": false, "error": { "code": "ERROR_CODE", "message": "...", "details": {} } }
    ```
servers:
  - url: http://localhost:3000
    description: Local development

# ─── Reusable components ─────────────────────────────────────────────────────
components:
  parameters:
    ProjectId:
      name: X-Project-ID
      in: header
      required: true
      description: Project identifier for multi-tenant isolation
      schema:
        type: string
        example: my-project
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - INVALID_TRANSITION
                - WIP_LIMIT_EXCEEDED
                - INVALID_STAGE
                - DEPENDENCY_CYCLE
                - DEPENDENCIES_NOT_MET
                - OUTPUT_COLLISION
                - CLAIM_CONFLICT
                - ITEM_CLAIMED
                - NOT_CLAIMED
                - CLAIM_MISMATCH
                - ITEM_NOT_FOUND
                - MISSION_NOT_FOUND
                - NO_ACTIVE_MISSION
                - INVALID_MISSION_STATE
                - PROJECT_NOT_FOUND
                - STAGE_NOT_FOUND
                - CONFLICT
                - UNAUTHORIZED
                - SERVER_ERROR
                - DATABASE_ERROR
            message:
              type: string
            details:
              description: Optional structured context for the error
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /api/learnings:
    parameters:
      - $ref: '#/components/parameters/ProjectId'

    post:
      operationId: captureLearning
      summary: Capture a RetroLearning row (per-mission dedupe)
      description: |
        Creates a RetroLearning row scoped to the requesting project, for the
        retro agent's Debrief and the corpus backfill. Dedupe is app-level and
        scoped per mission: a second call with the same (missionId, fingerprint)
        returns the existing row instead of creating a duplicate. A null
        missionId (backfill of stalled drafts) always inserts a new row.
      tags: [learnings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, severity, attributedAgent, targetSurface, pattern, fingerprint, title]
              properties:
                source:
                  type: string
                  description: Agent that surfaced the learning, e.g. 'stockwell', 'amy', 'lynch'
                severity:
                  type: string
                  description: "'low' | 'medium' | 'high' | 'critical'"
                attributedAgent:
                  type: string
                  description: Agent whose behavior the learning is about
                targetSurface:
                  type: string
                  description: File the learning would change, e.g. 'agents/ba.md'
                pattern:
                  type: string
                  description: Short pattern slug
                fingerprint:
                  type: string
                  description: Dedup key, computed in application code
                title:
                  type: string
                detail:
                  type: string
                  nullable: true
                missionId:
                  type: string
                  nullable: true
                  description: Null for backfill of stalled drafts; always inserted, never deduped
      responses:
        '201':
          description: New row created
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    required: [id]
                    properties:
                      id:
                        type: integer
        '200':
          description: Existing row returned (deduped on missionId + fingerprint)
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    required: [id]
                    properties:
                      id:
                        type: integer
        '400':
          description: VALIDATION_ERROR — missing required field, or missing/invalid X-Project-ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: |
            MISSION_NOT_FOUND — missionId is non-null but does not resolve to a
            mission owned by the requesting project (foreign or nonexistent;
            both cases return the same response so existence is not leaked).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/tuning/corroboration:
    parameters:
      - $ref: '#/components/parameters/ProjectId'

    get:
      operationId: getTuningCorroboration
      summary: Cross-mission corroboration signal for a fingerprint
      description: |
        Global — not scoped to the requesting project. Returns the
        distinct-mission count backing the `>= 3` corroboration gate.
      tags: [tuning]
      parameters:
        - name: fingerprint
          in: query
          required: true
          schema:
            type: string
          description: Fingerprint slug to check
      responses:
        '200':
          description: Corroboration data
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    required: [distinctMissions, corroborated]
                    properties:
                      distinctMissions:
                        type: integer
                      corroborated:
                        type: boolean
        '400':
          description: VALIDATION_ERROR — missing X-Project-ID or fingerprint query param
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/tuning/fingerprints/{slug}/defer:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: slug
        in: path
        required: true
        schema:
          type: string

    post:
      operationId: deferTuningFingerprint
      summary: Durable "not now" watermark for a fingerprint
      description: |
        No request body. Sets `Fingerprint.deferredAtMissions` to the
        fingerprint's current `distinctMissions`. The fingerprint reports
        `actionable: false` on `GET /api/tuning/candidates` until
        `distinctMissions` climbs DEFER_MARGIN past that watermark — this
        replaces the removed `reject`/`demote` verbs.
      tags: [tuning]
      responses:
        '200':
          description: Fingerprint deferred
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    required: [slug, deferredAtMissions, distinctMissions]
                    properties:
                      slug:
                        type: string
                      deferredAtMissions:
                        type: integer
                      distinctMissions:
                        type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: FINGERPRINT_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'

  # ── Missions ─────────────────────────────────────────────────────────────────
...